Refactor options to use OOP style of programming
This commits overhauls the options page from using the original
messy and inconsistent function type programming to Object-Oriented
Programming. Overall, making it easier to add and expand upon
in the future.

Overall Changes:
- The tabs data was converted into its own directory, where
a person can easily add, change, or remove an option
- Several new functions were created to split the behavior of the
original main function is to make it read in less than five lines.